Refactor and correct linting issues
General minor linting fixes.
Reflected resource name changes, refactored long lines into component statements, favoured string interpolation over concatenation.
